+++ packages/amplify-category-notifications/src/commands/notifications/status.ts
@@ -1,4 +1,4 @@
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { printer } from '@aws-amplify/amplify-prompts';
 import chalk from 'chalk';
 import { IChannelAvailability, INotificationsConfigStatus } from '../../channel-types';