+++ packages/amplify-util-uibuilder/src/event-handlers/handle-PostEnvAdd.ts
@@ -1,4 +1,4 @@
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { printer } from '@aws-amplify/amplify-prompts';
 import { run as cloneComponentsFromEnv } from '../commands/cloneComponentsFromEnv';